Add resize-helper script
this script can be used to resize the root fs partition. By default we can
create 'small' rootfs size, which can be installed on any board. On first boot
this script can find the root device, and extend the partition size and resize
the file system to make use of the entire eMMC.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>