Add a profile file to choose the correct UART for the Sensors board
The Arduino toolchain needs some help to know which UART to use when
running on a 96Board. This file adds two environmental variables to
configure both the IDE and command line Arduino tools.

JAVA_TOOL_OPTIONS adds the correct UART to the GUI IDE
MONITOR_PORT sets the correct UART for the Arduino.mk Makefile tools

Signed-off-by: Grant Likely <grant.likely@linaro.org>